1.7

+++ src/lesson3/task1/Loop.kt
@@ -208,7 +208,7 @@ fun isCoPrime(m: Int, n: Int): Boolean {
  * Использовать операции со строками в этой задаче запрещается.
  */
 fun revert(n: Int): Int {
-    var new = 0
+    var new = 0.0
     var c = n
     var k = 0
     while (c > 0) {
@@ -216,8 +216,7 @@ fun revert(n: Int): Int {
         new *= 10
         c /= 10
     }
-    return new / 10
-
+    return (new / 10).toInt()
 }
 
 /**
